refactor(tools): extract import-graph scanner
Both dependency gates — detect-dead-code and the upcoming
acyclic-dependencies gate — need the same notion of "production
source" and the same TS/ESM specifier resolution (a relative `.js`
specifier resolves to the `.ts` source on disk).

Move those primitives into ImportGraphScan.ts so the two gates stay
in lock-step instead of each carrying a drifting copy — the very
coupling the cycle gate exists to prevent. Behaviour-preserving: the
dead-code gate and its unit test pass unchanged.
